// src/SmartAbp.Vue/src/api/generated/models/SmartAbp_Application_Contracts_CodeGenerator_CqrsValidationResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SmartAbp_Application_Contracts_CodeGenerator_ValidationErrorDto } from './SmartAbp_Application_Contracts_CodeGenerator_ValidationErrorDto';
export type SmartAbp_Application_Contracts_CodeGenerator_CqrsValidationResultDto = {
    isValid?: boolean;
    errors?: Array<SmartAbp_Application_Contracts_CodeGenerator_ValidationErrorDto> | null;
};