// src/SmartAbp.Vue/src/api/generated/models/SmartAbp_CodeGenerator_Services_GeneratedCqrsSolutionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SmartAbp_CodeGenerator_Services_GeneratedCqrsSolutionDto = {
    moduleName?: string | null;
    files?: Record<string, string> | null;
    commandCount?: number;
    queryCount?: number;
    eventCount?: number;
    generatedAt?: string;
    sessionId?: string | null;
};