// src/SmartAbp.Vue/src/api/generated/models/SmartAbp_CodeGenerator_Services_V9_GeneratedModuleDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SmartAbp_CodeGenerator_Services_V9_GeneratedModuleDto = {
    moduleName?: string | null;
    generatedFiles?: Array<string> | null;
    generationReport?: string | null;
    sessionId?: string | null;
    success?: boolean;
    message?: string | null;
};